Move formulas to their own directory, use imperative style for formulas
