+++ server/src/formulas/index.ts
@@ -1,0 +1,18 @@
+import * as yup from 'yup';
+
+import { Formula } from "../types/attainable";
+import { CalculationResult, registerFormula, Status } from "../types/formulas";
+
+registerFormula(
+  Formula.Manual,
+  yup.object(),
+  // If no points have been input for a student, assume the attainment
+  // has been failed.
+  async (
+    _params: any,
+    _subGrades: Array<CalculationResult>,
+  ): Promise<CalculationResult> => {
+    return { status: Status.Fail, points: undefined };
+  },
+);
+
